Module webxr_api::session

Structs

Quitter

Session
An object that represents an XR session. This is owned by the content thread. https://www.w3.org/TR/webxr/#xrsession-interface

SessionBuilder
A type for building XR sessions

SessionId

SessionInit
https://immersive-web.github.io/webxr/#dictdef-xrsessioninit

SessionThread
For devices that want to do their own thread management, the SessionThread type is exposed.

Enums

EnvironmentBlendMode
https://immersive-web.github.io/webxr-ar-module/#xrenvironmentblendmode-enum

RenderState

SessionMode
https://www.w3.org/TR/webxr/#xrsessionmode-enum

SessionMsg

Statics

TIMEOUT

Traits

MainThreadSession
Devices that need to can run sessions on the main thread.