Struct webxr_api::session::SessionBuilder

pub struct SessionBuilder<'a, GL> {
    sessions: &'a mut Vec<Box<dyn MainThreadSession>>,
    frame_sender: Sender<Frame>,
    layer_grand_manager: LayerGrandManager<GL>,
    id: SessionId,
}

A type for building XR sessions

Fields
sessions: &'a mut Vec<Box<dyn MainThreadSession>>frame_sender: Sender<Frame>layer_grand_manager: LayerGrandManager<GL>id: SessionId
Implementations

impl<'a, GL: 'static> SessionBuilder<'a, GL>

pub fn id(&self) -> SessionId

pub(crate) fn new( sessions: &'a mut Vec<Box<dyn MainThreadSession>>, frame_sender: Sender<Frame>, layer_grand_manager: LayerGrandManager<GL>, id: SessionId, ) -> Self

pub fn spawn<Device, Factory>(self, factory: Factory) -> Result<Session, Error>
where Factory: 'static + FnOnce(LayerGrandManager<GL>) -> Result<Device, Error> + Send, Device: DeviceAPI,

For devices which are happy to hand over thread management to webxr.

pub fn run_on_main_thread<Device, Factory>( self, factory: Factory, ) -> Result<Session, Error>
where Factory: 'static + FnOnce(LayerGrandManager<GL>) -> Result<Device, Error>, Device: DeviceAPI,

For devices that need to run on the main thread.