Function webxr_api::util::fov_to_projection_matrix

pub fn fov_to_projection_matrix<T, U>(
    left: f32, 
    right: f32, 
    top: f32, 
    bottom: f32, 
    clip_planes: ClipPlanes
) -> Transform3D<f32, T, U>

Construct a projection matrix given the four angles from the center for the faces of the viewing frustum