Module webxr_api::view

This crate uses euclid's typed units, and exposes different coordinate spaces.

Structs

View
For each eye, the transform from the viewer to that eye, its projection onto its display, and its display viewport. For stereo displays, we have a View<LeftEye> and a View<RightEye>. For mono displays, we hagve a View<Viewer> (where the transform is the identity). https://immersive-web.github.io/webxr/#xrview

Enums

Display
The normalized device coordinate space, where the display is from (-1,-1) to (1,1).

Floor
The coordinate space of the floor https://immersive-web.github.io/webxr/#dom-xrreferencespacetype-local-floor

Input
The coordinate space of an input device

LeftEye
The coordinate space of the left eye https://immersive-web.github.io/webxr/#dom-xreye-left

Native
The native 3D coordinate space of the device This is not part of the webvr specification.

RightEye
The coordinate space of the right eye https://immersive-web.github.io/webxr/#dom-xreye-right

Viewer
The coordinate space of the viewer https://immersive-web.github.io/webxr/#dom-xrreferencespacetype-viewer

Viewport
The unnormalized device coordinate space, where the display is from (0,0) to (w,h), measured in pixels.

Views
Whether a device is mono or stereo, and the views it supports.

Constants

_IMPL_DESERIALIZE_FOR_Display

_IMPL_DESERIALIZE_FOR_Floor

_IMPL_DESERIALIZE_FOR_Input

_IMPL_DESERIALIZE_FOR_LeftEye

_IMPL_DESERIALIZE_FOR_Native

_IMPL_DESERIALIZE_FOR_RightEye

_IMPL_DESERIALIZE_FOR_View

_IMPL_DESERIALIZE_FOR_Viewer

_IMPL_DESERIALIZE_FOR_Viewport

_IMPL_DESERIALIZE_FOR_Views

_IMPL_SERIALIZE_FOR_Display

_IMPL_SERIALIZE_FOR_Floor

_IMPL_SERIALIZE_FOR_Input

_IMPL_SERIALIZE_FOR_LeftEye

_IMPL_SERIALIZE_FOR_Native

_IMPL_SERIALIZE_FOR_RightEye

_IMPL_SERIALIZE_FOR_View

_IMPL_SERIALIZE_FOR_Viewer

_IMPL_SERIALIZE_FOR_Viewport

_IMPL_SERIALIZE_FOR_Views