Module webxr_api::view

This crate uses euclid's typed units, and exposes different coordinate spaces.

Structs

SomeEye
Pattern-match on eyes

View
For each eye, the pose of that eye, its projection onto its display. For stereo displays, we have a View<LeftEye> and a View<RightEye>. For mono displays, we hagve a View<Viewer> https://immersive-web.github.io/webxr/#xrview

Viewports
A list of viewports per-eye in the order of fields in Views.

Enums

Capture
The coordinate space of a secondary capture view

CubeBack
The coordinate space of the back frustrum of a cubemap

CubeBottom
The coordinate space of the bottom frustrum of a cubemap

CubeLeft
The coordinate space of the left frustrum of a cubemap

CubeRight
The coordinate space of the right frustrum of a cubemap

CubeTop
The coordinate space of the top frustrum of a cubemap

Display
The normalized device coordinate space, where the display is from (-1,-1) to (1,1).

Floor
The coordinate space of the floor https://immersive-web.github.io/webxr/#dom-xrreferencespacetype-local-floor

Input
The coordinate space of an input device

LeftEye
The coordinate space of the left eye https://immersive-web.github.io/webxr/#dom-xreye-left

Native
The native 3D coordinate space of the device This is not part of the webvr specification.

RightEye
The coordinate space of the right eye https://immersive-web.github.io/webxr/#dom-xreye-right

Viewer
The coordinate space of the viewer https://immersive-web.github.io/webxr/#dom-xrreferencespacetype-viewer

Viewport
The unnormalized device coordinate space, where the display is from (0,0) to (w,h), measured in pixels.

Views
Whether a device is mono or stereo, and the views it supports.

Constants

CUBE_BACK

CUBE_BOTTOM

CUBE_LEFT

CUBE_RIGHT

CUBE_TOP

LEFT_EYE

RIGHT_EYE

VIEWER